refactor: Add MutableSignedRational.
