+++ Sources/SternBroctTreeSwift/Core/Rational16.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 /// A rational type for value semantics.
-public struct Rational16 : SignedRational {
+public struct Rational16 : MutableSignedRational {
 
     public typealias Number = Int16
 
@@ -36,166 +36,6 @@ public struct Rational16 : SignedRational {
     private init(_ numerator: Int16, _ denominator: Int16) {
         self.numerator = numerator
         self.denominator = denominator
-    }
-
-    /// Returns a value wether this value can simplify or not.
-    ///
-    /// - Complexity: O(log n) where n is digits of the given `denominator`.
-    public var canSimplify: Bool {
-        let commonFactor = gcd(numerator, denominator)
-        return commonFactor != 1 && commonFactor != -1
-    }
-
-    /// Returns a new simplified rational.
-    ///
-    /// Returns new value when the numerator and the denominator have common devider except for ± 1,
-    ///
-    ///     let new = Rational(fraction: "3/9").simplified
-    ///     // new.description is 1/3.
-    ///
-    /// otherwise always returns self.
-    ///
-    ///     let new = Rational(fraction: "3/10").simplified
-    ///     // new.description is 3/10.
-    ///
-    /// - Note:
-    /// `Reduce` is term used to reduce numerics by gcm, but  `simplified` execute sign inversion of the numerator and the denominator in addition.
-    public func simplified() -> Rational16 {
-
-        var x = self
-        x.simplify()
-
-        return x
-    }
-
-    /// Mutate this value to a simplified rational.
-    public mutating func simplify() {
-
-        let commonFactor = gcd(numerator, denominator)
-
-        var numerator = self.numerator
-        var denominator = self.denominator
-
-        // Tricky: avoid overflow from (INT32_MIN / -1)
-        // The range of Int16 is 2147483647 ~ -2147483648. Because positive range includes zero, these ranges are
-        // asymmetry, so an error will occur when trying to multiplied Int16.min by -1. It causes overflow.
-        if commonFactor != -1 || (numerator != Int16.min && denominator != Int16.min) {
-            numerator /= commonFactor
-            denominator /= commonFactor
-        }
-
-        // prevent negative denominator, but do not negate the smallest value that would produce overflow
-        if denominator < 0 && numerator != Int16.min && denominator != Int16.min {
-            numerator *= -1
-            denominator *= -1
-        }
-
-        self.numerator = numerator
-        self.denominator = denominator
-    }
-
-    // MARK: - Arithmetic Operations
-
-    public func addingReportingOverflow(_ other: Rational16) -> (partialValue: Rational16, overflow: Bool) {
-
-        var x = self
-        var y = other
-        var xNumeratorYDenominator, yNumeratorYDenominator, numerator, denominator: Int16!
-        var isROverflowed, isSOverflowed, isNumeratorOverflowed, isDenominatorOverflowed: Bool!
-
-        var retry = true
-        while retry {
-
-            (xNumeratorYDenominator, isROverflowed) = x.numerator.multipliedReportingOverflow(by: y.denominator)
-            (yNumeratorYDenominator, isSOverflowed) = y.numerator.multipliedReportingOverflow(by: x.denominator)
-            (numerator,isNumeratorOverflowed) = xNumeratorYDenominator.addingReportingOverflow(yNumeratorYDenominator)
-            (denominator, isDenominatorOverflowed) = x.denominator.multipliedReportingOverflow(by: y.denominator)
-
-            // 1. Go into a branch if r,s, numerator or denominator is overflowed.
-            // 2. If they can simplify, do it. If not, throw overflow error.
-            // 3. Continue the steps to resolve overflow completely.
-            if isROverflowed || isSOverflowed || isNumeratorOverflowed || isDenominatorOverflowed {
-
-                // overflow in intermediate value
-                if !x.canSimplify && !y.canSimplify {
-
-                    // neither fraction could reduce, cannot proceed
-                    return (Rational16(numerator, denominator), true)
-                }
-
-                x.simplify()
-                y.simplify()
-
-                // the fraction(s) reduced, good for one more retry
-                retry = true
-            } else {
-                retry = false
-            }
-
-        }
-
-        return (Rational16(numerator, denominator).simplified(), false)
-
-    }
-
-    public func subtractingReportingOverflow(_ other: Rational16) -> (partialValue: Rational16, overflow: Bool) {
-        return addingReportingOverflow(-other)
-    }
-
-    public func multipliedReportingOverflow(by other: Rational16) -> (partialValue: Rational16, overflow: Bool) {
-        var x = self
-        var y = other
-        var numerator, denominator: Int16!
-        var isNumeratorOverflowed, isDenominatorOverflowed: Bool!
-
-        var retry = true
-        while retry {
-
-            (numerator, isNumeratorOverflowed) = x.numerator.multipliedReportingOverflow(by: y.numerator)
-            (denominator, isDenominatorOverflowed) = x.denominator.multipliedReportingOverflow(by: y.denominator)
-
-            // 1. Go into a branch if x or y is overflowed.
-            // 2. If they can simplify, do it. If not, throw outOfRange(overflow) error.
-            // 3. Continue the steps to resolve overflow completely.
-            if isNumeratorOverflowed || isDenominatorOverflowed {
-
-                // overflow in intermediate value
-                if !x.canSimplify && !y.canSimplify {
-
-                    // neither fraction could reduce, cannot proceed
-                    return (Rational16(numerator, denominator), true)
-                }
-
-                x.simplify()
-                y.simplify()
-
-                // the fraction(s) reduced, good for one more retry
-                retry = true
-            } else {
-                retry = false
-            }
-        }
-
-        return (Rational16(numerator, denominator).simplified(), false)
-    }
-
-    public func dividedReportingOverflow(by other: Rational16) -> (partialValue: Rational16, overflow: Bool) {
-        return multipliedReportingOverflow(by: Rational16(other.denominator, other.numerator))
-    }
-
-    public mutating func negate() {
-        if numerator == Int16.min {
-
-            let simplified = self.simplified()
-
-            // check again
-            if simplified.numerator == Int16.min {
-
-                // denominator can't be MIN too or fraction would have previosly simplifed to 1/1.
-                self = Rational16(simplified.numerator, simplified.denominator * -1)
-            }
-        }
-        self = Rational16(numerator * -1, denominator)
     }
 
     public func backwardingMatrixSequence() -> [Matrix2x2] {
@@ -228,20 +68,6 @@ public struct Rational16 : SignedRational {
 
         return box.flatMap({ $0 })
     }
-
-    /// Returns a mediant from two fractions.
-    public static func mediant(left: Self, right: Self) throws -> Self {
-
-        let (numeratorAddingResult, numeratorAddingOverflow) = left.numerator.addingReportingOverflow(right.numerator)
-        let (denominatorAddingResult, denominatorAddingOverflow) = left.denominator.addingReportingOverflow(right.denominator)
-
-        if numeratorAddingOverflow || denominatorAddingOverflow {
-            throw RationalError.overflow(lhs: left, rhs: right)
-        }
-
-        return Self(numerator: numeratorAddingResult,denominator: denominatorAddingResult)
-    }
-
 
 }
